+++ libs/core/lib/command.ts
@@ -1,5 +1,0 @@
-export interface Command {
-  readonly type: string;
-  readonly by: string;
-  readonly timestamp: string;
-}