+++ libs/core/lib/handler-map.ts
@@ -1,3 +1,0 @@
-import { CommandHandler } from './command-handler';
-
-export type HandlerMap = Record<string, CommandHandler>;